pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "asyncio>=4.0.0",
    "datetime>=5.5",
    "discord-py>=2.5.2",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "logging>=0.4.9.6",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
